+++ src/test/groovy/ru/vyarus/java/generics/resolver/SpecialMapsTest.groovy
@@ -1,6 +1,5 @@
 package ru.vyarus.java.generics.resolver
 
-import ru.vyarus.java.generics.resolver.context.container.ParameterizedTypeImpl
 import ru.vyarus.java.generics.resolver.error.UnknownGenericException
 import ru.vyarus.java.generics.resolver.support.Base2
 import ru.vyarus.java.generics.resolver.util.TypeToStringUtils
@@ -8,6 +7,8 @@ import ru.vyarus.java.generics.resolver.util.map.EmptyGenericsMap
 import ru.vyarus.java.generics.resolver.util.map.IgnoreGenericsMap
 import ru.vyarus.java.generics.resolver.util.map.PrintableGenericsMap
 import spock.lang.Specification
+
+import static ru.vyarus.java.generics.resolver.util.type.TypeFactory.*
 
 import java.lang.reflect.Type
 
@@ -33,13 +34,13 @@ class SpecialMapsTest extends Specification {
     def "Check ignoring map behaviour"() {
 
         when: "verification case"
-        TypeToStringUtils.toStringType(new ParameterizedTypeImpl(Base2, Base2.getTypeParameters()), ["K": String])
+        TypeToStringUtils.toStringType(param(Base2, Base2.getTypeParameters()), ["K": String])
         then: "error"
         def ex = thrown(UnknownGenericException)
         ex.genericName == "P"
 
         when: "using ignoring map"
-        def res = TypeToStringUtils.toStringType(new ParameterizedTypeImpl(Base2, Base2.getTypeParameters()),
+        def res = TypeToStringUtils.toStringType(param(Base2, Base2.getTypeParameters()),
                 new IgnoreGenericsMap(["K": String]))
         then:
         res == "Base2<String, Object>"
@@ -58,13 +59,13 @@ class SpecialMapsTest extends Specification {
     def "Check printable generic names"() {
 
         when: "verification case"
-        TypeToStringUtils.toStringType(new ParameterizedTypeImpl(Base2, Base2.getTypeParameters()), ["K": String])
+        TypeToStringUtils.toStringType(param(Base2, Base2.getTypeParameters()), ["K": String])
         then: "error"
         def ex = thrown(UnknownGenericException)
         ex.genericName == "P"
 
         when: "using ignoring map"
-        def res = TypeToStringUtils.toStringType(new ParameterizedTypeImpl(Base2, Base2.getTypeParameters()),
+        def res = TypeToStringUtils.toStringType(param(Base2, Base2.getTypeParameters()),
                 new PrintableGenericsMap(["K": String]))
         then:
         res == "Base2<String, P>"